+++ src/main/java/com/pixulse/infx/mixin/world/entity/MobCategoryCapMixin.java
@@ -1,5 +1,6 @@
 package com.pixulse.infx.mixin.world.entity;
 
+import com.pixulse.infx.config.InfXConfig;
 import net.minecraft.world.entity.MobCategory;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -11,7 +12,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class MobCategoryCapMixin {
     @Inject(method = "getMaxInstancesPerChunk", at = @At("HEAD"), cancellable = true)
     private void infx$miteMonsterCap(CallbackInfoReturnable<Integer> cir) {
-        if ((Object) this == MobCategory.MONSTER) {
+        if (InfXConfig.INSTANCE.mobs.enabled.getValue() && (Object) this == MobCategory.MONSTER) {
             cir.setReturnValue(50);
         }
     }